Move events to common mod code Fixes no version reported on integrated server
