+++ src/main/java/jobicade/betterhud/network/MessageNotifyClientHandler.java
@@ -8,6 +8,8 @@ import net.minecraftforge.fml.common.network.simpleimpl.MessageContext;
 public class MessageNotifyClientHandler implements IMessageHandler<MessageVersion, IMessage> {
 	@Override
 	public IMessage onMessage(MessageVersion message, MessageContext context) {
+		BetterHud.getLogger().info("Server reported version " + message.version.getVersionString());
+
 		BetterHud.getProxy().setServerVersion(message.version);
 		return null;
 	}
