+++ src/main/java/jobicade/betterhud/proxy/ClientProxy.java
@@ -1,14 +1,14 @@
 package jobicade.betterhud.proxy;
 
-import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
-import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
-import net.minecraftforge.fml.common.network.FMLNetworkEvent.ClientDisconnectionFromServerEvent;
 import net.minecraftforge.fml.common.versioning.ArtifactVersion;
 import net.minecraftforge.fml.common.versioning.DefaultArtifactVersion;
 
-@EventBusSubscriber
 public class ClientProxy implements HudSidedProxy {
-    private ArtifactVersion serverVersion = new DefaultArtifactVersion("0.0");
+    private ArtifactVersion serverVersion;
+
+    public ClientProxy() {
+        setServerVersion(null);
+    }
 
     @Override
     public ArtifactVersion getServerVersion() {
@@ -17,11 +17,10 @@ public class ClientProxy implements HudSidedProxy {
 
     @Override
     public void setServerVersion(ArtifactVersion version) {
-        serverVersion = version;
+        if (version == null) {
+            serverVersion = new DefaultArtifactVersion("version");
+        } else {
+            serverVersion = version;
+        }
     }
-
-	@SubscribeEvent
-	public void onPlayerDisconnected(ClientDisconnectionFromServerEvent e) {
-		serverVersion = new DefaultArtifactVersion("0.0");
-	}
 }
